Improved MinGW32 support and added Win64 support

+++ src/jpeg-6/jmorecfg.h
@@ -1,3 +1,5 @@
+#include <stdint.h>
+
 /*
  * jmorecfg.h
  *
@@ -149,7 +151,7 @@ typedef unsigned short UINT16;
 typedef unsigned int UINT16;
 #endif /* HAVE_UNSIGNED_SHORT */
 
-typedef long INT32;
+typedef int32_t INT32;
 
 /* INT16 must hold at least the values -32768..32767. */
